past_tense_future_tense/requirements.txt
openai
langchain
streamlit
python-dotenv
langchain_openai
